requirements.txt
numpy
pandas
Pillow
scikit-learn
timm
torch
torchvision
tqdm
